[gen] use int64_t id/weight for larger graphs

+++ src/benchmark.h
@@ -28,8 +28,8 @@ Various helper functions to ease writing of kernels
 
 
 // Default type signatures for commonly used types
-typedef int32_t NodeID;
-typedef int32_t WeightT;
+typedef int64_t NodeID;
+typedef int64_t WeightT;
 typedef NodeWeight<NodeID, WeightT> WNode;
 
 typedef CSRGraph<NodeID> Graph;
